# openapi.yaml
openapi: 3.0.3
info:
  title: UnityAssetsManager API
  version: 1.5.1
  description: API locale pour explorer et exporter des assets Unity (CSV/SQLite). Le port HTTP par défaut est 5003 et peut être modifié dans config/config.json.
servers:
  - url: http://localhost:5003
paths:
  /api/data:
    get:
      summary: Liste paginée des données assets
      parameters:
        - in: query
          name: draw
          schema:
            type: integer
        - in: query
          name: start
          schema:
            type: integer
        - in: query
          name: length
          schema:
            type: integer
        - in: query
          name: search[value]
          schema:
            type: string
        - in: query
          name: search[regex]
          schema:
            type: boolean
        - in: query
          name: profile
          schema:
            type: string
          description: Nom du profil à appliquer (charge filter_stack et alias_map du profil)
        - in: query
          name: filter_stack
          schema:
            type: string
          description: Stack de filtres sérialisée en JSON
        - in: query
          name: alias_map
          schema:
            type: string
          description: Map d'alias de colonnes sérialisée en JSON
        - in: query
          name: filter_invalid_assets
          schema:
            type: boolean
          description: Exclure les assets invalides (slug+url manquants)
        - in: query
          name: order[0][column]
          schema:
            type: integer
          description: Index de la colonne de tri
        - in: query
          name: order[0][dir]
          schema:
            type: string
            enum: [asc, desc]
          description: Direction du tri
      responses:
        '200':
          description: Données paginées
          content:
            application/json:
              schema:
                type: object
                properties:
                  draw:
                    type: integer
                  recordsTotal:
                    type: integer
                  recordsFiltered:
                    type: integer
                  data:
                    type: array
                    items:
                      type: object
        '404':
          description: Aucune donnee chargee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /api/columns:
    get:
      summary: Liste des colonnes disponibles
      responses:
        '200':
          description: Colonnes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '404':
          description: Aucune donnee chargee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
...
  /api/profiles:
    get:
      summary: Liste des profils
      responses:
        '200':
          description: Profils
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    post:
      summary: Sauvegarder un profil
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileSaveRequest'
      responses:
        '200':
          description: Profil sauvegardé
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        '400':
          description: Requete invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '500':
          description: Erreur interne
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /api/profiles/{name}:
    get:
      summary: Charger un profil
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Profil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '404':
          description: Profil absent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
    delete:
      summary: Supprimer un profil
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Profil supprimé
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        '404':
          description: Profil absent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /api/export:
    post:
      summary: Exporter les données filtrées
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Fichier exporté (binaire)
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Erreur de requête ou de template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '500':
          description: Erreur export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /api/batch-export:
    post:
      summary: Exporter les données en mode headless vers un fichier local
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchExportRequest'
      responses:
        '200':
          description: Export batch écrit sur disque
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  file:
                    type: string
                  format:
                    type: string
                  count:
                    type: integer
        '400':
          description: Requête batch invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '500':
          description: Erreur de batch export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /api/reload:
    post:
      summary: Recharger les données depuis la source
      responses:
        '200':
          description: Rechargement effectué
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        '500':
          description: Erreur lors du rechargement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /api/stats:
    get:
      summary: Statistiques sur la source courante
      responses:
        '200':
          description: Statistiques
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Aucune donnee chargee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
...
  /api/setup:
    post:
      summary: Configurer la source de données
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupRequest'
      responses:
        '200':
          description: Configuration appliquée
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        '400':
          description: Paramètres de configuration invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '500':
          description: Erreur de configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /api/config:
    get:
      summary: Lire la configuration runtime publique
      responses:
        '200':
          description: Configuration courante
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  config:
                    $ref: '#/components/schemas/PublicRuntimeConfig'
    post:
      summary: Mettre a jour la configuration runtime autorisee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigUpdateRequest'
      responses:
        '200':
          description: Configuration mise a jour
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  config:
                    $ref: '#/components/schemas/PublicRuntimeConfig'
        '400':
          description: Payload de configuration invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '500':
          description: Erreur interne
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
  /api/test-path:
    post:
      summary: Vérifier un chemin source CSV/SQLite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
      responses:
        '200':
          description: Resultat du test
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
                  is_dir:
                    type: boolean
                  type:
                    type: string
                  source_type:
                    type: string
                  rows:
                    type: integer
                  cols:
                    type: integer
                  tables:
                    type: array
                    items:
                      type: string
                  warning:
                    type: string
components:
  schemas:
    ApiErrorEnvelope:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
      required:
        - error
...
    SetupRequest:
      type: object
      properties:
        data_path:
          type: string
        db_table:
          type: string
        show_parser_warnings:
          type: boolean
          description: Afficher les avertissements d'analyse CSV
    ProfileSaveRequest:
      type: object
      properties:
        name:
          type: string
        columns:
          type: array
          items:
            type: string
        filter_stack:
          type: array
          items:
            type: object
        column_aliases:
          type: object
          additionalProperties: true
        description:
          type: string
      required:
        - name
    ProfileResponse:
      type: object
      properties:
        name:
          type: string
        columns:
          type: array
          items:
            type: string
        description:
          type: string
        lastUpdated:
          type: string
        filter_stack:
          type: array
          items:
            type: object
        column_aliases:
          type: object
          additionalProperties: true
        column_profile:
          type: object
    ExportRequest:
      type: object
      properties:
        template:
          type: string
        profile:
          type: string
          description: Nom du profil à appliquer (charge filter_stack et alias_map du profil)
        filter_stack:
          type: array
          items:
            type: object
        alias_map:
          type: object
          additionalProperties:
            type: string
        filter_invalid_assets:
          type: boolean
          description: Applique le filtre d'assets invalides uniquement si true (export interactif).
    BatchExportRequest:
      allOf:
        - $ref: '#/components/schemas/ExportRequest'
        - type: object
          properties:
            output_dir:
              type: string
            file_name:
              type: string
            output_path:
              type: string
            profile:
              type: string
...